+++ docs/.vuepress/configs/sidebar/index.ts
@@ -1,0 +1,4 @@
+import type { SidebarConfig } from 'vuepress-theme-mix'
+
+export const en: SidebarConfig = {
+}